refactor: update navbar import in home and dashboard pages
